complete rework
