+++ .gitignore
@@ -1,4 +1,5 @@
-/node_modules/
-/public/build/
-/.vscode
-/venv
+*/node_modules/
+*/build/
+*/.vscode
+/.idea
+*.env
