+++ api/tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "compilerOptions": {
+        "forceConsistentCasingInFileNames": true,
+        "module": "commonjs",
+        "esModuleInterop": true,
+        "outDir": "./build",
+        "rootDir": "./src",
+        "target": "es6",
+        "skipLibCheck": true,
+        "strict": true
+    }
+}
