+++ configs/.env.example
@@ -1,0 +1,7 @@
+POSTGRES_USER=
+POSTGRES_PASSWORD=
+POSTGRES_DB=
+POSTGRES_HOST=
+DATABASE_URL=
+HOST=
+API_PORT=
