+++ database/tables.sql
@@ -1,0 +1,9 @@
+DROP TABLE IF EXISTS countries;
+
+CREATE TABLE countries (
+  id SERIAL PRIMARY KEY,
+  name VARCHAR(100) UNIQUE NOT NULL,
+  lat REAL NOT NULL,
+  lng REAL NOT NULL,
+  correct BOOLEAN DEFAULT FALSE
+);
