--- guessera-api/requirements.txt
@@ -1,82 +1,0 @@
-anyio==3.5.0
-appdirs==1.4.4
-black==21.12b0
-btrfsutil==5.16
-CacheControl==0.12.6
-certifi==2021.10.8
-cffi==1.15.0
-chardet==4.0.0
-charset-normalizer==2.0.10
-click==8.0.3
-colorama==0.4.4
-contextlib2==0.6.0.post1
-cryptography==36.0.1
-defusedxml==0.7.1
-deprecation==2.1.0
-distlib==0.3.4
-distro==1.6.0
-dnspython==2.1.0
-Flask==2.0.2
-Flask-Cors==3.0.10
-future==0.18.2
-Glances==3.2.4
-gotrue==0.5.0
-greenlet==1.1.2
-grpcio==1.43.0
-h11==0.12.0
-html5lib==1.1
-httpcore==0.14.5
-httpx==0.21.3
-idna==3.3
-importlib-metadata==4.8.1
-isc==2.0
-itsdangerous==2.0.1
-Jinja2==3.0.3
-lensfun==0.3.95
-Markdown==3.3.6
-MarkupSafe==2.0.1
-meld==3.20.4
-more-itertools==8.10.0
-msgpack==1.0.3
-mypy-extensions==0.4.3
-ordered-set==4.0.2
-packaging==20.9
-pathspec==0.9.0
-pep517==0.12.0
-platformdirs==2.4.1
-ply==3.11
-postgrest-py==0.8.1
-progress==1.6
-psutil==5.9.0
-pwquality==1.4.4
-pycairo==1.20.1
-pycparser==2.21
-pydantic==1.9.0
-PyGObject==3.42.0
-pynvim==0.4.3
-pyOpenSSL==21.0.0
-pyparsing==2.4.7
-python-dateutil==2.8.2
-python-dotenv==0.19.2
-pyxdg==0.27
-realtime==0.0.4
-Reflector==2021.7.8.19.52.52
-requests==2.26.0
-resolvelib==0.5.5
-retrying==1.3.3
-rfc3986==1.5.0
-shtab==1.5.3
-six==1.16.0
-sniffio==1.2.0
-supabase==0.3.2
-team==1.0
-termcolor==1.1.0
-tldr==3.0.0
-toml==0.10.2
-tomli==1.2.3
-typing-extensions==4.0.1
-urllib3==1.26.7
-webencodings==0.5.1
-websockets==9.1
-Werkzeug==2.0.2
-zipp==3.6.0
