--- guessera-api/sql/tables.sql
@@ -1,10 +1,0 @@
--- SQLite
-DROP TABLE IF EXISTS countries;
-
-CREATE TABLE countries (
-  id INTEGER PRIMARY KEY AUTOINCREMENT,
-  name TEXT UNIQUE NOT NULL,
-  lat REAL NOT NULL,
-  lng REAL NOT NULL,
-  correct BOOLEAN DEFAULT 0
-);
