--- package.json
@@ -1,29 +1,0 @@
-{
-	"name": "svelte-app",
-	"version": "1.0.0",
-	"private": true,
-	"scripts": {
-		"build": "rollup -c",
-		"dev": "rollup -c -w",
-		"start": "sirv public --no-clear",
-		"format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. ."
-	},
-	"devDependencies": {
-		"@rollup/plugin-commonjs": "^17.0.0",
-		"@rollup/plugin-node-resolve": "^11.0.0",
-		"@types/three": "^0.136.1",
-		"prettier": "^2.5.1",
-		"rollup": "^2.3.4",
-		"rollup-plugin-css-only": "^3.1.0",
-		"rollup-plugin-glslify": "^1.2.1",
-		"rollup-plugin-livereload": "^2.0.0",
-		"rollup-plugin-svelte": "^7.1.0",
-		"rollup-plugin-terser": "^7.0.0",
-		"svelte": "^3.0.0"
-	},
-	"dependencies": {
-		"@supabase/supabase-js": "^1.29.3",
-		"sirv-cli": "^2.0.0",
-		"three": "^0.136.0"
-	}
-}
