--- public/global.css
@@ -1,41 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Outfit:wght@300;700;900&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap');
-
-:root {
-	/* Typefaces | hf = headfont */
-	--xlarge-hf: 3.75rem; /* 60px */
-	--large-hf: 3rem; /* 48px */
-	--med-hf: 2.25rem; /* 36px */
-	--small-hf: 1.875rem; /* 30px */
-	--base-f: 1rem; /* 16px */
-	--xbase-f: 1.125rem; /* 18px */
-	/* Colour */
-	/* catppuccin color palette */
-	--rosewater: #f5dfdb;
-	--flamingo: #f2cece;
-	--mauve: #ddb5f2;
-	--pink: #f5c2e7;
-	--maroon: #e8a1ae;
-	--red: #f28dab;
-	--peach: #f8bd96;
-	--yellow: #fae3b2;
-	--green: #aae9b3;
-	--teal: #b5e8e1;
-	--blue: #98cefb;
-	--sky: #89dceb;
-	--lavender: #c7c9ff;
-	--black0: #120f1f;
-	--black1: #1b1924;
-	--black2: #1f1f2e;
-	--black3: #312e42;
-	--black4: #565167;
-	--gray0: #6e6c7f;
-	--gray1: #988ba2;
-	--gray2: #c2b9c6;
-	--white: #d8dfee;
-}
-
-* {
-	padding: 0;
-	margin: 0;
-	box-sizing: border-box;
-}
