--- public/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<meta name="viewport" content="width=device-width,initial-scale=1" />
-
-		<title>Guessera</title>
-
-		<link rel="icon" type="image/png" href="/favicon.svg" />
-		<link rel="stylesheet" href="/global.css" />
-		<link rel="stylesheet" href="/build/bundle.css" />
-
-		<script defer src="/build/bundle.js"></script>
-	</head>
-
-	<body></body>
-</html>
