+++ web/src/global.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="svelte" />
